add bookmarkable settings to vault web page
Introduce a GET form and add submit button to submit all inputs and thus
create a bookmarkable URL. On page load a javascript parses the query
parameters back into the form input values. Removed the name from the
passphrase inputs so they won't be in the URL.

Added some missing semicolons as it's either automatic semicolon insertion
or not. ;-)

Made some CSS selectors that have a 'width' specified using 'inline-block'
instead of 'inline'.